Enemy loot, treasure chest and persistent data
Made enemies able to drop loot. Made interactable objects, in this case a treasure chest, and made it possible for the player to interact with the s button. Lastly a persistence handler scene was made, to be given to objects, like the treasure chest, so it can remember player interactions with it.

Also did some bug fixing and some small code cleanup to align it more with GDscript's style guidelines.